Rename Common / Generator files to be okit prefixed not oci
