+++ visualiser/facades/ociBackend.py
@@ -15,7 +15,7 @@ __module__ = "ociBackends"
 
 import oci
 
-from common.ociLogging import getLogger
+from common.okitLogging import getLogger
 from facades.ociConnection import OCILoadBalancerConnection
 
 # Configure logging
