+++ visualiser/facades/ociBlockStorageVolumes.py
@@ -15,7 +15,7 @@ __module__ = "ociBlockStorageVolumes"
 
 import oci
 
-from common.ociLogging import getLogger
+from common.okitLogging import getLogger
 from facades.ociConnection import OCIBlockStorageVolumeConnection
 
 # Configure logging
