+++ visualiser/facades/ociBootVolumeAttachment.py
@@ -15,7 +15,7 @@ __module__ = "ociBootVolumeAttachments"
 
 import oci
 
-from common.ociLogging import getLogger
+from common.okitLogging import getLogger
 from facades.ociConnection import OCIComputeConnection
 from facades.ociBootVolumes import OCIBootVolumes
