+++ visualiser/facades/ociCompartment.py
@@ -15,7 +15,7 @@ __module__ = "ociCompartment"
 
 import oci
 
-from common.ociLogging import getLogger
+from common.okitLogging import getLogger
 from facades.ociConnection import OCIIdentityConnection
 from facades.ociInstance import OCIInstanceVnics
 from facades.ociInstance import OCIInstances
