+++ visualiser/facades/ociImage.py
@@ -16,8 +16,8 @@ __module__ = "ociImage"
 
 import oci
 
-from common.ociLogging import getLogger
-from common.ociCommon import logJson
+from common.okitLogging import getLogger
+from common.okitCommon import logJson
 from facades.ociConnection import OCIComputeConnection
 
 # Configure logging
