+++ visualiser/facades/ociRegion.py
@@ -15,8 +15,8 @@ __module__ = "ociRegion"
 
 import oci
 
-from common.ociCommon import logJson
-from common.ociLogging import getLogger
+from common.okitCommon import logJson
+from common.okitLogging import getLogger
 from facades.ociConnection import OCIIdentityConnection
 
 # Configure logging
