+++ visualiser/facades/ociResourceManager.py
@@ -17,8 +17,8 @@ import base64
 import oci
 import time
 
-from common.ociLogging import getLogger
-from common.ociCommon import logJson
+from common.okitLogging import getLogger
+from common.okitCommon import logJson
 from facades.ociConnection import OCIResourceManagerConnection
 
 # Configure logging
