+++ visualiser/facades/ociVnic.py
@@ -15,8 +15,8 @@ __module__ = "ociVnic"
 
 import oci
 
-from common.ociLogging import getLogger
-from common.ociCommon import logJson
+from common.okitLogging import getLogger
+from common.okitCommon import logJson
 from facades.ociConnection import OCIVirtualNetworkConnection
 
 # Configure logging
