+++ visualiser/facades/ociVnicAttachement.py
@@ -15,8 +15,8 @@ __module__ = "ociVnicAttachments"
 
 import oci
 
-from common.ociLogging import getLogger
-from common.ociCommon import logJson
+from common.okitLogging import getLogger
+from common.okitCommon import logJson
 from facades.ociConnection import OCIComputeConnection
 from facades.ociVnic import OCIVnics
