+++ visualiser/generators/okitTerraform11Generator.py
@@ -14,9 +14,9 @@ __module__ = "ociTerraform11Generator"
 
 import os
 
-from common.ociCommon import writeTerraformFile
-from common.ociLogging import getLogger
-from generators.ociGenerator import OCIGenerator
+from common.okitCommon import writeTerraformFile
+from common.okitLogging import getLogger
+from generators.okitGenerator import OCIGenerator
 
 # Configure logging
 logger = getLogger()
