+++ visualiser/model/okitCostEstimator.py
@@ -12,7 +12,7 @@ __version__ = "1.0.0"
 __module__ = "ociPricing"
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
 
-from common.ociLogging import getLogger
+from common.okitLogging import getLogger
 
 # Configure logging
 logger = getLogger()
